Add support for generating lockfile with rpmtree
This change introduces generation of the lockfile from the rpmtree
command.

Following the previous patterns with the other bazel files, this adds
some helper methods to bazel.go and creates a handler object to wrap
this up for rpmtree.

Worth noting is that to make this work we also have to tweak AddTree()
to take the configuration name because the generated repos are
different when using the lockfile.

This change leaves the MODULE.bazel configuration up to the user as
there may be some additional configuration for bazeldnf that is beyond
the scope of what rpmtree is doing and we don't want to interfere with
that.